Start refactoring DexTypeList access (2/2)
Summary:
Try to deprecate the `get_type_list` `Deque` access as much as possible. The container already had `size`, `begin`, `end` and `at`, so was readily available to access directly and in for-each loops.

Only code that needs to mutate a type list still retains access to the renamed `get_type_list_internal`. This should, however, really be moved to something (mostly) handled in `DexTypeList` itself.

In the process of this, avoid making `Deque` copies in some places that really didn't need to.

Reviewed By: NTillmann

Differential Revision: D31659049

fbshipit-source-id: 7ad4fa325ecbf97071e25414b4182ce72f7f6ee5