introduce IOPCODE_INIT_CLASS instruction
Summary:
The `init-class` instruction has a  `type` property, similar to `const-class` and `new-instance`, but has no register arguments or results. Its presence indicates that the static initializer of the mentioned class must run at that point in time.

This new instruction can be inserted when `new-instance`, `invoke-static`, and `sget`/`sput` instructions are eliminated to keep the static initialization semantics.

In subsequent diffs...
- a pass will be introduced to lower this instruction to an `sget` of an existing or a new field,
- existing passes will get audited and modified to insert `init-class` instructions as needed,
- the InterDex pass needs to reserve as many field refs as might be needed for existing and potentially needed intra- and inter-dex `init-class` instructions
- semantically redundant `init-class` instructions will get eliminated,
- we'll refine the `clinit_may_have_side_effects` analysis reduce the number of classes for which `init-class` instructions are needed

Reviewed By: beicy

Differential Revision: D31129740

fbshipit-source-id: f545bbb58065cac090420fd7cc4acb02ddc564a2