Introducing AppModuleUsagePass
Summary: Introducing the new pass `AppModuleUsagePass` to open source. The redex `AppModuleUsagePass` analyzes cross store/app module references. An app module is a DexStore that is not the root store. When the root store or another app module makes a reference to code in a different DexStore (which is not the root store)  that is a cross store reference. `AppModuleUsagePass` considers cross store references not accompanied by the `UsesAppModule` annotation a "violation". By default the pass fails on a violation, but it can be configured to continue with `crash_with_violations` in the config, if it continues a list of all violations is produced at `redex-app-module-annotation-violations.csv`. If the pass does not fail it also generates `redex-app-module-usage.csv` and `redex-app-module-count.csv` which describe all cross store references and how often they are referenced.

Reviewed By: yuxuanchen1997

Differential Revision: D31770876

fbshipit-source-id: be20fe2403eda2b9663d131c64bf41d8d01d1814