Only track class-loading instructions in `get_types_usage`
Summary:
Add a mode to class merging that only tracks class-loading instructions for type references. All other types will report no references.

This is in an effort to only group classes that are actually loaded together, not just used (which would imply earlier load).

In general, a full interprocedural static analysis seems necesssary to make this precise. However, the mostly-transitively-closed nature of the input type set, coming from profiling, means this is OK as-is.

Reviewed By: NTillmann

Differential Revision: D31292890

fbshipit-source-id: 4429dc6f9339cd550fbf94beaa47adb61e6ae8f1