Intern the deobfuscated string
Summary:
Because of aliasing we already create a `DexString` anyways, so use that to avoid duplicating the string data.

Change the interface a bit for now to make it more explicit that deobfuscated names might not be set.

Provide `get_deobfuscated_name_or_empty` as a backwards solution.

Lowers profiled memory (frontend, average of 10):
| Before |  After |
| ------ | ------ |
| 7950.5 | 7160.8 |

Reviewed By: NTillmann

Differential Revision: D31722843

fbshipit-source-id: 758b8f7f8f93fb656d2cd1e2d4f4c96f3a58d5c5