Make target store name configurable
Summary:
In particular, configuring it as "empty" results in the separated out dexes being placed at the end of the root store.
Together with a configuration change, this is a behavior preserving diff.
(The idea is that this can be used for an early A/B test to assess the impact of just separating out the dead classes into their own dex, without super-packing that dex.)

Differential Revision: D31914781

fbshipit-source-id: bed6b582314c0ca5000e5f67498ed3cabab6d90e