relaxing checks for certain Kotlin marker methods
Summary:
We have observed that the Kotlin compiler injects invocations to a certain marker in a way that causes imbalanced monitor stack. We choose to ignore that here, as it seems benign in the observed cases due to the Android verifier not doing exhaustive checks.

Also fixes the pretty-printed monitor count transition in the error message to remove ignored edges, and show the initial state.

Reviewed By: agampe

Differential Revision: D32037358

fbshipit-source-id: 0f8bee38574fe1d6b6e5c7560c5cd5f30032bf80