add Kotlin coroutines test for ClassMerging
Summary:
As in the title. This exposes the pattern we want to optimize in the next diff. Currently, there will be 6 constructors generated, that all look very similar: P461624188.

This stack aims to weaken the argument types representing instance fields that are merged by class merging under new fields of type Object. This way, the existing merging logic of ClassMerging for constructors will be able to dedup more code.

Reviewed By: thezhangwei

Differential Revision: D30698621

fbshipit-source-id: 1649db3b3a1838de74188703e1d83df2f85b9e0a