AppModuleUsagePass allowlist prefixes and any store
Summary: In the allowlist/prexisting violation file, multiple entrypoints can be added to the allowlist with the same allowed stores by providing a prefix and ending with an asterisk. An entrypoint or prefix can also now be allowed to use any module without annotation by replacing the store name list with an asterisk.

Reviewed By: yuxuanchen1997

Differential Revision: D32328782

fbshipit-source-id: 98bc80e10b6ea37eb8d972c24f2269bbd0270449