Fix setting source-map in redex.py when --lldb is passed
Summary:
An lldb command needs to be run at the start of a debugging session, so lldb knows where sources are.

The way the argument is being passed in redex.py when --lldb is used, using the
python `quote` function for escaping, is causing the command to be passed to
lldb as a single token wrapped in single quotes, which does not work:

```
(lldb) 'settings set target.source-map "." "/home/jimmycleary/fbsource"'
error: command 'settings' did not recognize 'settings  set target.source-map "." "/home/jimmycleary/fbsource"' as valid (subcommand might be invalid).
```

Solution is to escape the `src_dir` with `quote`, not the whole command.

Reviewed By: agampe

Differential Revision: D32436689

fbshipit-source-id: 41e104f9015f8e06b3c613f214df19496a1df848