Super detailed AppModuleUsage file
Summary:
Adds the file 'redex-app-module-verbose-details.txt' which details every instance when the AppModuleUsagePass finds a cross app module reference. Example output: P467202450
This output isn't particularly useful as a whole but being able to search for individual entries is useful to see which classes are being used to cause an AppModuleUsage reference.

Reviewed By: yuxuanchen1997

Differential Revision: D32319239

fbshipit-source-id: adac617114a6c8d80a1c40142bc50e4e9ffc7b41