change DexString and some string parsing to use string_view
Summary: A `string_view` is just a modern representation of a `const char*` and its associated length. A `string_view` isn't required to be null terminated. It's especially useful for parsing because substring operations are essentially free, avoiding frequent allocations. This diff also changes the keying of s_string_map, which may or may not have a perf implication.

Reviewed By: agampe

Differential Revision: D32779569

fbshipit-source-id: cbb144b79c27e35d43c49b1afec1b18d337ec9e7