Allow complex insertion
Summary:
Allow complex insertion, that is, a mix of `IRInstruction`, `SourceBlock` and `DexPosition`, in `ControlFlowGraph` and `CFGMutation`. This is done by widening the accepted iterated types in the `insert_(before|after)` overloads. As long as the types can be coerced to `cfg::ControlFlowGraph::InsertVariant = std::variant<IRInstruction*, std::unique_ptr<SourceBlock>, std::unique_ptr<DexPosition>>`. The compiler should be optimizing away the variant overhead for the older invocations.

The `CFGMutation` functions have new names as otherwise the overloads are ambiguous. Handling is not as natural for the variant overload as initializer lists do not allow move semantics which are required for the `unique_ptr`s and thus the `variant` - the vector needs to be explicitly constructed.

The `CFGMutation` is also restricted, in that different methods cannot be mixed. This is because otherwise the whole code becomes complex (having to make copies vs forwarding etc), and in any way serialization of different requests isn't clear.

Added tests.

Reviewed By: NTillmann

Differential Revision: D33004641

fbshipit-source-id: b62b04f2c466c59e12e4a220c1a88b6f3c91f164