extending SignedConstantDomain with intervals
Summary:
Before, `SignedConstantDomain`, was the reduced product of...
- sign_domain::Domain and
- ConstantDomain

This diff refines it by adding a
- SmallNonNegativeIntervalDomain
which is an `IntervalDomain` over `int32_t` (2^31-1 and -2^31 represent bounds that extend to infinity).
(See D33012933 for a greatly simplification implementation that more directly shows how the domains interact.)

To avoid an exploding abstract state in our abstract interpretation fixpoint computations, we do not refine operators like `+c` to shift or enlarge the interval.

This domain refinement by itself allows constant-propagation to do more, e.g. when deciding branching conditions.
It also enables in later diffs... 1) better reasoning about infeasible switch default cases, 2) refining values used in branches, and 3) tracking api-level values.

Future work:

While fully functional, the resulting internal representation of `SignedConstantDomain` is suboptimal:
- It's 72 bytes big, which consists of a vtable pointer, plus 16 bytes for the sign_domain::Domain (vtable pointer + payload), 32 bytes for the constant domain (vtable pointer plus vtable pointer plus AbstractValueKind plus actual 64-bit payload) plus 16 additional bytes for the SmallNonNegativeIntervalDomain (vtable pointer plus two 32-bit values (lower/upper bounds)).
- The `reduce_product` function is a bit involved due to the need to propagate across all subdomains

Really, all that should be needed to express our payload should be...
- 24 bytes: 8 bytes for a vtable pointer, plus 16 bytes for two 64-bit values (lower/upper bounds), where special values such as bottom and not-zero (and possibly more... almost all other not-X values? or maybe some small non-consecutive sets of small numbers? the sub-stack under D32964409 shows that there's potential) can be encoded in values where the upper bound is less than the lower bound. Or we could even throw in another 8 byte value to round it up to 32 bytes --- still less than the current 56 bytes.

From that, the `sign_domain::Domain` and `ConstantDomain` that are queried and used everywhere can be derived, and we don't even need to deal with the (minor) complication of bounds anymore.

Reviewed By: amnn

Differential Revision: D32857796

fbshipit-source-id: 9904d33d4478b716738178cb59d37485693e992a