rewire infeasible switch gotos
Summary:
If we know that the default case of a switch is infeasible, rewrite the branches with the most common target block to be the new effective default case.

We know that the default case is infeasible when the (permissible interval of the) selector value is completely covered by branches with matching case keys.

Reviewed By: agampe

Differential Revision: D32859203

fbshipit-source-id: 2cf311cec5dc28ea380681474fdb3dabcb0531c8