compute definitely assigned ifields, don't force-include 0 in their WPS field partition
Summary:
"Definitely assigned" instance fields are those for which we can guarantee that any reading happens after a write.
For those, observable values are comprised only of all ever written values, and we don't have to take into account the default value 0.
This allows us to tighten the whole-program-state field partitions, and consequently do more constant-propagation optimizations in IPCP.

Reviewed By: beicy

Differential Revision: D33138911

fbshipit-source-id: 1edc0ca1d30bcd8d7ac1f305c53a2ce0000877f9