actually don't consider cold-start classes
Summary:
We always properly used method-profiles to identify which methods are "hot". However, this information is not always reliable before the time the InstrumentationPass runs.
As another line of defense, we also used to consider all methods of all cold-start classes as "hot". The way this was done would only work after the InterDexPass runs. However, the PartialApplicationPass runs before...

So, in this diff, before InterDexPass runs, we don't look at the `is_perf_sensitive` property (which is only set by InterDexPass), but instead we look at the configured `coldstart_classes` instead.

Reviewed By: thezhangwei

Differential Revision: D33288656

fbshipit-source-id: 9ec196ae55ed84eaf63895e3e49f7b034fcc8266