Introduce builder classes for resources.arsc types
Summary: Make some rudimentary builder style APIs for rebuilding the various string pools in the resource table. There are obviously a lot of shortcomings and things to improve here, but the goal is to get the format implemented correctly and clean up inefficiencies later (and expand it to the other types in the file later too).

Reviewed By: ssj933

Differential Revision: D32691498

fbshipit-source-id: 452156ab801529d1a4bd9f31a99d227fd424b331