Fix too much sorting
Summary:
The ordering was done locally for each virtual hierarchy. As the order is deterministic and independent, cache the result.

A different approach would be to cull the profiled methods to the virtual hierarchy and then sort. This seems more effort and is not a clear-case winner, so avoid the complexity.

Reviewed By: NTillmann

Differential Revision: D33521267

fbshipit-source-id: 00dcbff8fcbb0b8ed72b824f29bb282fb09e10ce