Add aggregated stats for all config models in ClassMergingPass
Summary:
Having the total pass wise stats are useful sometimes to spot regressions.

Currently, if we drop a model completely for some reason, it's hardly visible in the stats. Because the stats of the affected model are gone in the base to diff comparison. Having a set of total stats tracking the aggregated numbers for all the models merged in the pass helps to show the overall trend/change of the pass.

Reviewed By: Feng23

Differential Revision: D33777435

fbshipit-source-id: b70aefac4caa430218fde61e8d1b5bc2833e22e6