Don't introduce spurious moves
Summary:
Usually, CSE is "forwarding" previously computed values by inserting moves (one to store the previously computed value in a temp register, and then a later one to move it to the register where that value is expected when it's reused).
However, when forwarding const-literal instructions, we don't want to do that, as moves are "typed" (wide/object/-), and this can lead to problems with const-0 instructions, when they get used with different "types" along the same execution path.

Thus, special code already existed that avoided inserting the second move, and instead, there it would just clone the const-instructions.
However, the first move would still get inserted. Always typed as (wide/-), but never as object. And this can cause the const-0 value to be "used" as both (object) and (-) along the same execution path, which is illegal. The idea was that this shouldn't really matter, as local-dce would immediately remove those useless moves (that target a temp register that is never read).

But before local-dce typically runs, copy-propagation runs. And copy-propagation picks up the differently typed aliases for those constants along the same execution path, and does the wrong thing. (It caused type violations in the next diff up the stack.)

This is getting fixed here, by simply not inserting those spurious first moves that are never used. We also avoid allocating a dummy register for them.

Reviewed By: beicy

Differential Revision: D33723259

fbshipit-source-id: 5cc2add45e70a7d6805e6091b5cf84b1a8c3ca2c