Manually dedupe in exit-block instrumentation
Summary:
Do not rely on DedupBlocksPass to clean up exit instrumentation:
* We usually run register allocation before dedupe (to attempt to give dedupe more opportunities). With multiple exit blocks, the instrumentation becomes expensive because of many renamed registers.
* Register allocation does make no attempt (yet) to use the same register for exit dataflow (return or throw values). It may not produce a syntactical form that can be deduped.

Manual deduping of the instrumentation is made more complicated by catch handlers. As such, for each ordered list of catch blocks, a new instrumentation is generated. If and when Redex analyzes the feasibility of throw edges better, this may be reduced.

Reviewed By: minjang

Differential Revision: D33733397

fbshipit-source-id: 0be5f0576973be29cda3668a21d150ecc0617a9b