compute is_init_or_clinit from original, not copy
Summary:
The ObjectEscapeAnalysis creates method clones to operate on (so that we can throw away the result if it becomes to big or we run into inlining problems, and to not interfere with any other ongoing concurrent operations). The cloned methods gets a different names to not clash with the original one. However, this would change "<init>" to "<init>$redex_stack_allocated", which would no longer pass the `method::is_init` check, causing CSE to perform wrong transformations on final fields.

To avoid that issue, we use the new `shrink_code` method, where we can pass the metadata explicitly, providing correct information about whether the method at hand is a constructor.

Reviewed By: agampe

Differential Revision: D33933981

fbshipit-source-id: 1d96bea0e4eeb6db1ba525f274b4b11449b72ab2