Transitive closure for interdex order
Summary:
When loading the interdex order from the cold-start classes, optionally transitively close the list. That is, if a class is in the list, ensure that superclasses and interfaces - that are in the app and not in the primary - are ordered before. This reflects that those other classes are necessary when loading the class in question.

This overrides the choice made in the list as at this point we *know* that the other classes are actually necessary. That is, pulling classes into the coldstart dex, for example, does not make sense here.

NOTE: It would be better to close the class-order before any passes, such that performance decisions could be made upon a "correct" list earlier. However, as we also optimize away classes and the coldstart order is more reflective of the state at the end of the optimization, this could preclude too many size optimizations.

Reviewed By: NTillmann

Differential Revision: D34253324

fbshipit-source-id: 7bd36beb4c9922bb719ec3df521dd3da924efa1e