generalization and refinement of check-cast workaround
Summary:
This is a generalization (in the sense that we now already do it, not only after the (final) RegAlloc pass has run), and a refinement (in the sense that we only break those aliases that need breaking) of D26237484 (https://github.com/facebook/redex/commit/a59d854d40dd58b777cb4504e89a31665815bb74), as copy-prop must never extend live-ranges into throw handlers that happen to overlap with check-cast destination registers.
That's because (at least as things are today) we can never recover from it, even when RegAlloc runs again.

(The sentence "The register allocator ensures that the register is not live before the check-cast so that this works." on the original diff is not quite accurate --- the register allocator doesn't ensure it, as it won't break up a problematic pre-existing live-ranges. Maybe it should, but as things are today it doesn't.)

Reviewed By: agampe

Differential Revision: D34264212

fbshipit-source-id: bc326ca2e4fe7cfba0c95c15222bec72c6f0b4a6