order blocks
Summary:
As is well known for linear-scan register allocations, the particular order in which basic blocks are linearized doesn't matter for the correctness of the algorithm.

By ordering blocks in a way similar to how dedup-blocks will visit them, and by ordering in a way that prefers short lifetime intervals, we can improve linear scan register allocation and dedup-block efficiency.

Reviewed By: agampe

Differential Revision: D33994235

fbshipit-source-id: 557458dae7806f46d23e5b8614a57a36c09b2d59