work around check-cast IR limitation
Summary:
The move-result-pseudo-object associated with a check-cast must not have the same dest register as the src(0) of the check cast, if that dest register is live-in to any catch handler of the check-cast.

Here, we add auxiliary live-ranges that make the check-casts move-result-pseudo-object's dest register appear live-in to catch handler target blocks (if it isn't already live).

This is required to work around a quirk in our IR representation of check-cast instruction, which doesn't match actual DEX instruction.

Reviewed By: agampe

Differential Revision: D34133789

fbshipit-source-id: dfc6ac23d559b62b53b3476962873e911db71fda