prefer making max-case-key not min-case-key into goto target
Summary:
When the default-branch of a switch is infeasible, we can remove one switch case. Everything else being equal, we used to remove the switch case key with the lowest case key. This is changing here to be the largest key, which tends to produce smaller code.

This also changes the unnecessary O(nlog n) set construction into something linear.

Reviewed By: Feng23

Differential Revision: D34221167

fbshipit-source-id: 12b0b0719bf75b60cf73db1490f418abb047e3b7