maintain transitive resolved dependencies
Summary:
And use them to address a TODO in the code. Replaces linear string-based dependency scan with a constant-time lookup (*), and potentially reduces number of illegal refs in case transitive dependencies were not passed to Redex.

(*) I doubt there's a noticeable actual speed-up, given that dependency sets are quite small. The main reason for this diff is that better dependency information is needed for the next diff up the stack.

Reviewed By: beicy

Differential Revision: D34286287

fbshipit-source-id: 0c7ffc28e284bf4f175226d134b153827d022d24