docs: add social banner in support of Ukraine (#680)
Summary:
Our mission at [Meta Open Source](https://opensource.facebook.com/) is to empower communities through open source, and we believe that it means building a welcoming and safe environment for all. As a part of this work, we are adding this banner in support for Ukraine during this crisis.

![image](https://user-images.githubusercontent.com/12485205/156666237-7628902f-fc95-4786-a50f-c0f9ca12a80d.png)

Pull Request resolved: https://github.com/facebook/redex/pull/680

Test Plan:
Imported from GitHub, without a `Test Plan:` line.

**Static Docs Preview: redex**
|[Full Site](https://our.intern.facebook.com/intern/staticdocs/eph/D34627056/V2/redex/)|

|**Modified Pages**|

Reviewed By: ssj933

Differential Revision: D34627056

Pulled By: yuxuanchen1997

fbshipit-source-id: 797c86de172f48df37731003871d885b2d43f133