docs: add GH button in support of Ukraine (#17)
Summary:
Our mission at Meta Open Source is to empower communities through open source, and we believe that it means building a welcoming and safe environment for all. As a part of this work, we are adding this banner in support for Ukraine during this crisis.

X-link: https://github.com/facebook/SPARTA/pull/17

Reviewed By: arnaudvenet

Differential Revision: D34635651

Pulled By: dmitryvinn-fb

fbshipit-source-id: af430dab6470fc92da76a8b982a39461a4a3a946