Don't use stale values
Summary:
D32859203 (https://github.com/facebook/redex/commit/1aeeaaba7e44d1a88c60f1e538b037c25a53e7d6) introduced an optimization to "rewire infeasible switch gotos".
It also introduced a bug where it could remove a relevant switch case.
In the code, the core of the bug is that the reference to `goto_edge->target()` of a stale (to-be-deleted) `goto_edge` is used, where it really should consider the new ("rewired") goto-target.

This is getting fixed here.

Reviewed By: agampe

Differential Revision: D34705547

fbshipit-source-id: 97eec3e9577d2f4433998ff879379c8bd3bfe92e