fixing ResolveMethodRefs mutation
Summary:
In practice, the code was already operating on editable cfgs. However, it seemingly issued mutation operations on the (otherwise inactive) `code` itself.
This went unnoticed, even though it actually mutated an editable block's entries. However, the custom allocator diff up the stack detects this issues.
So we clean it up here, using the CFGMutator instead.

This is a behavior-preserving change.

Reviewed By: thezhangwei

Differential Revision: D34915903

fbshipit-source-id: 3b6474f66c93f3980ffbde0407bd7bdfbaf6dc99