Exclude mergeable class with virtual calls
Summary:
VerticalMerging tries to merge superclass and subclass pairs when only one can be instantiated.
This diff tries to fix a corner case when a subclass is uninstantiable but virtual calls on the class do exist. The virtual calls are possibly unreachable at runtime, but we need to make the code well-formed.

Reviewed By: thezhangwei, ssj933

Differential Revision: D34975992

fbshipit-source-id: 8c50a05e25249b6bb3345ac05d35ac9886e82135