global cost function
Summary:
Instead of making all net-savings computations and decisions locally (for one root method at a time), we now postpone decisions that depend on families of root methods all creating objects of the same common types, and then later make a global net-savings computation and decision.

This diff only affects the cost function, not the actual logic of doing the object escape analysis and code transformation.

Reviewed By: beicy

Differential Revision: D34571869

fbshipit-source-id: fea09ee49e77e07359a420c5b7d7620254432e71