expanding constructor parameters
Summary:
When the ObjectEscapeAnalysis can proof that a newly created object doesn't escape, it attempts to inline all method invocations interacting with the object. However, when an invoked method is a constructor, then inlining may only be possible when the constructed object (the first constructor parameter) itself doesn't escape. Otherwise, we would have a problem keeping the code verifiable.

When another object parameter (not the first constructor parameter) doesn't escape, and the invoked constructor only reads fields of the non-escaping object (and in particular, doesn't write to fields), then we can get around the issue by replacing the parameter with a sequence of field value parameters, and adjust the constructor code to replace field reads with references to the field value parameters.

This is what happens here.

We don't actually mutate any constructor, but add additional ones, speculating that the original constructor will most likely become unreachable (we don't account for the additional constructor in our cost function).

There are a few details we have to be careful about (there could be conflicts with constructor parameter types, deobfuscated names, or previously removed methods; or we could add too many parameters). We only actually concretize the new constructors at the end, after we know exactly which ones have been selected via the (global) cost function, and erase all other temporarily created method refs that didn't ultimately get selected.

Reviewed By: beicy

Differential Revision: D34603227

fbshipit-source-id: 14d09af6a171a078fa3e07fa0b1cef27bc0b0261