Getting closer to O(change size) instead of O(code size)
Summary:
When flushing a CFGMutation, we used to iterate over all instructions, of a potentially huge method, even if we were touching just a single thing.

This is getting fixed here: We only iterate over affected blocks, and we can stop when we are done in any particular block.

This also makes dealing with move-results more predictable:
- Anchoring at a move-result-any instruction whose primary instruction is being replaced should not be done as it will not be honored.
- Actually, we used to only not honor it if the move-result-any appeared immediate adjacent to its primary instruction; if instead it sat as the first instruction in an earlier block, which it theoretically (but never in practice) could have, then it actually would have been honored, which is just inconsistent and weird.
- We now consistently don't honor it, and issue a WARNING into the trace stream when this could cause actual inconsistencies (i.e. it's not the case that both are removals).

Thus, (in practice), this is a behavior-preserving change.

Reviewed By: amnn

Differential Revision: D34810303

fbshipit-source-id: 09e2c355e57aede145a65e8c8ac45a41776a1d0f