Delete resource inlining
Summary: This was never deployed anywhere. The reason was the logic never obeyed the types of values that an attribute was declared to have (i.e. you can define an XML attribute which is allowed to hold a reference but NOT a color, for example and inlining would cause a bug).

Reviewed By: ssj933

Differential Revision: D34360891

fbshipit-source-id: 9c20b6c03249665c97f75155d1abac0d3c6a01f3