File size assert
Summary:
I have stumbled over this several times in the past, and always forgot about it. `write_serialized_data()` will not work properly if the serialized bytes are larger than the file itself.

In a shipping app, this likely will never happen because we shrink the file, but for test cases this poses an annoying challenge.

Reviewed By: ssj933

Differential Revision: D34469983

fbshipit-source-id: 40ab73ffe743febd6b86034c96b0b0e04fa152c6