Fix silly mistake from D34836968
Summary: This code only works when the output vector has nothing in it, which is actually all the time. However it's not actually right since who knows what callers could want to write the data to in theory, so make sure to set the header's size to the actual difference in how much the vector has grown.

Reviewed By: ssj933

Differential Revision: D35164129

fbshipit-source-id: c18f302a0ebedba7eb5190a6559c9cd60456511e