promote D35111828 & D35146468 to staging/stable
Summary: As in the title. The implementations in unstable/staging and stable are different for the method orderer, as we wish to APK A/B test the changes incrementally. However, this diff changes the orderer to take into account method profiles data, which is what we want to test in 361.

Reviewed By: NTillmann

Differential Revision: D35181616

fbshipit-source-id: c3079a5aaa426e9a041885c5b8b8e2821ebc8091