make SourceBlock smaller
Summary:
Instead of a resizable (but immutable in practice) `std::vector` that is 24 bytes on x64,  we can get away by using 4 bytes of otherwise wasted padding space to record the size, and a simple `std::unique_ptr<...[]>`. This reduces the size of a `SourceBlock` from 48 to 32 bytes.

This is a behavior-preserving change.

Reviewed By: ssj933

Differential Revision: D35130951

fbshipit-source-id: 6ff9245b6ccea185e2142b065e94154d41871363