TableParser base class for visiting/collecting arsc data
Summary:
PackageStringRefCollector has some generic functionality that would be useful for other kinds of traversals. So, pull some of its guts out to a super class, and define another sub class that stores entry data in a convenient fashion for lookups based on a certain config.

This will be used to migrate functionality off the AOSP forked logic.

Reviewed By: ssj933

Differential Revision: D35040128

fbshipit-source-id: 4a4295dcdb5f4509b0506f9efa055b8fb0ca2dfe