Move is_resource_file definition closer to its callers
Summary:
The existing `is_resource_file()` doesn't really know what it's being used for, and therefore has to guess whether or not what type of file path its given, whether this is from an .aab file or from the resource table, etc.

I propose we delete this file and make more targeted versions of this function to do only what is required at the call site.

Reviewed By: ssj933

Differential Revision: D35226826

fbshipit-source-id: fc3d6caa49d182d56756ac3b45c2749cd4721f6f