sweep members of sweeped classes
Summary:
When an entire class got "sweeped", i.e. simply removed from the dex, we used to not sweep the methods/fields in it. This is getting fixed here. In particular, sweeping methods frees the the contained code in them, which is quite significant.
We still don't delete the actual field/methods pointers in case any pass holds on to them.

This is a behavior-preserving change.

Reviewed By: ssj933

Differential Revision: D35134417

fbshipit-source-id: 74ddfa2e2e7428252680dd6d7c29020936fb7a16