Internalize "location" of classes
Summary:
This addresses a 5-year old TODO:
```
class DexClass {
...
  const std::string m_location; // TODO: string interning
```

Our solution here is a new `DexLocation` type, which is uniquely constructed and stored by the RedexContext just as `DexString` and other `Dex...` types.
(I am also sneaking a "store" name in there, which will be used in the next diff, and is the real reason I am doing this change now.)

This is a behavior preserving change.

Reviewed By: beicy

Differential Revision: D35202484

fbshipit-source-id: 45b2dbd628ec3936df05c9b4770f2c09226f320f