Make ResStringPoolBuilder more convenient
Summary:
Make it accept data in a variety of forms, and let it internally convert to the desired char width based on the flags. Simplify some other stuff with this.

As noted in the comments, this thing still expects callers to format strings properly as described in https://source.android.com/devices/tech/dalvik/dex-format#mutf-8. For our purposes this is easy; data truly originating from our tooling will be simple ascii style strings for type names, file paths, etc etc (user data in the input app will be copied as is so we don't really need the full blown encoder logic as far as I can tell).

Reviewed By: ssj933

Differential Revision: D35237322

fbshipit-source-id: 9477c8612cfa81567de21383b8ab83eca75b4906