store DexString payload in unique_ptr
Summary:
This is a first step to insure that we are no longer dependent on having an `std::string` backing of `DexString`s.

This reduces the size of each `DexString` from 64 bytes to 16 bytes, and moves the entire payload on the heap, whereas before the `std::string` small-string-optimization might have "inlined" some string data. The combined effect seems to be a slight reduction in memory usage. But the main purpose of this diff is to prepare for the next diff in the stack.

This is a behavior-preserving change.

Reviewed By: yuxuanchen1997

Differential Revision: D35066046

fbshipit-source-id: 4725a9d4f0bb8a3d72890c93b7dbee59863ce5aa