storing raw string payloads in dense string buffers
Summary:
With jemalloc, allocation sizes are rounded up to fit into buckets. This creates waste. In this diff, we are attacking that overhead of `DexString` payloads. We are switching over to storing the payloads in dense pre-allocated buffers that we neatly fill up, instead of individually allocating string payloads.

The buffers are stored in containers managed by a `ConcurrentStringStorage` facility that enables sharded and practically lock-free allocations (on x86) with extremely little waste. A key insight is that the container/buffer in which a payload ends up doesn't have to be deterministic / isn't tied to its content, and we can freely pick whatever container is most convenient for the current thread, allowing for many container to be actively extended in times of high contention, while occasionally sorting containers by remaining space to minimize waste. We partition strings into "small" strings (length<111) that are stored in pre-allocated buffers that are 16KB each, "medium" strings (111<=length<2000) in 64KB buffers, and all other "large" strings (there are few) sill get their own allocation.

Reviewed By: agampe

Differential Revision: D35067604

fbshipit-source-id: a7152f481f76fc1a4657fbf7a4dcc35c45841143