Do not inline large escaping callees
Summary:
Context:

The escaping call inlining loop is very aggressive. It tries to inline every call if inliner can handle it. However, doing so  could be ominous in some cases, since we could replicate too many instruction for a saving on meta data.

Previous we added an iteration cap that limit the number of iterations we run to remove a Builder. This limits how many levels do we chase the call chain in order to inline all possible escaping calls.

Here we simply add a size threshold on the escaping callees.

The change includes:
- In the escaping call inlining loop, if a callee is too large, stop inlining and give up
- For now, the too large threshold is a global constant `140` against the `sum_opcode_sizes()` on `IRCode`.

Reviewed By: Feng23

Differential Revision: D35449199

fbshipit-source-id: 720ba1f8bc2b915def3f794190e471a757439b17