Add support for the `includedescriptorclasses` proguard keep modifier
Summary:
Proguard defines a modifier, `includedescriptorclasses`, intended to propagate a `keep` rule to the args and return value (if a method) or type (if a field) of the member that a keep rule refers to. This is particularly useful for preserving classes used as args across JNI boundaries.

This diff adds support for this modifier, as well as some elementary tests. These verify that when used, we do not rename the classes of fields or method args.

Reviewed By: agampe

Differential Revision: D35494598

fbshipit-source-id: d842b85b18c98a1fe1081f66a2de25cc53e54c7e