Prevent outlining from DeepData analysis methods
Summary:
I was doing some local testing of a change to the DeepData format, expanding
call count to 32-bit to avoid the current overflow issue, and noticed that code
was being outlined from `onMethodExit` overloads. In the DeepData profile,
there were 3313431 calls in total, and 1904284 were to one of two of these
outlinees, or about 57% of all calls.

`onMethodExit` is also obviously extremely hot in dyna builds, since it's
unconditionally called by almost every method. Ordinarily, the outliner does
not outline from hot methods, but since we don't profile `onMethodExit` itself,
the outliner sees no reason not to outline from it.

I did not observe this reproing in any actual CI builds or BSB (with stable or
unstable), so it seems that something about my jest+dyna build of Messenger
with slight modifications to `onMethodExit` triggered it.

However, since this **can** happen, and it's not likely to be performant or
safe (in this case we were lucky that we don't place `onMethodExit` in
single-block methods, and that the outlinees are single block.  Otherwise, we'd
infinitely recurse into `onMethodExit`.), I've gone ahead and fixed it.

This change prevents this from happening by setting the `no_optimizations` flag
for analysis methods during `InstrumentPass::eval_pass`. This flag is cleared
before the shrinker runs at the end of `InstrumentPass` to prevent pessimizing
these very hot functions (and reset afterwards, just in case).

Reviewed By: agampe

Differential Revision: D35502712

fbshipit-source-id: bd9e596f914c6b573a62475c43c1d26b66450d85